+++ codeanalyzer/core.py
@@ -965,7 +965,6 @@ class Codeanalyzer:
                 skip_tests=self.skip_tests,
                 shard=self.options.pycg_shard,
                 shard_ceiling=self.options.pycg_shard_ceiling,
-                shard_timeout=self.options.pycg_shard_timeout,
                 shard_strategy=self.options.pycg_shard_strategy,
                 max_iter=self.options.pycg_max_iter,
                 using_ray=self.using_ray,