+++ codeanalyzer/options/options.py
@@ -62,7 +62,6 @@ class AnalysisOptions:
     verbosity: int = 0
     pycg_shard: bool = False
     pycg_shard_ceiling: int = 100
-    pycg_shard_timeout: int = 120
     pycg_shard_strategy: ShardStrategy = ShardStrategy.JEDI
     pycg_max_iter: int = 50
     entrypoint_rules: Tuple[Path, ...] = ()